# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "Complete E-Commerce API with Xendit Payment Integration and Automated Webhook Processing",
        "contact": {
            "name": "API Support",
            "email": "support@ecommerce.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ecommerce-api-test-production.up.railway.app",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/webhooks/payment": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Process payment webhook from Xendit",
                "operationId": "26a75d59b652507161f56bc0d509ab34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "external_id",
                                    "status",
                                    "amount"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "inv_test_123"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "example": "ORD-1705312800-1234"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "PENDING",
                                            "PAID",
                                            "EXPIRED",
                                            "FAILED"
                                        ],
                                        "example": "PAID"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000000
                                    },
                                    "payer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "customer@example.com"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "BANK_TRANSFER"
                                    },
                                    "paid_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-01-15T10:30:00.000Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook processed successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "processed": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "payment_id": {
                                                    "type": "string",
                                                    "example": "inv_test_123"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "PAID"
                                                },
                                                "order_status": {
                                                    "type": "string",
                                                    "example": "paid"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid webhook data"
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "API Access Key",
                "name": "X-ACCESS-KEY",
                "in": "header"
            },
            "secretKey": {
                "type": "apiKey",
                "description": "API Secret Key",
                "name": "X-SECRET-KEY",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}